Add a change log to the documentation.
A customer asked us to provide them with a list of what changes (in more
detail) from release to release. This is probably a good idea to
maintain. I'm sure it will be a common request. Will also help us track
our own progress on the API, and when changes were made.